portfolio evaluation parseFloat bug fix. Sale middleware refactor
